elete bet (#24)

+++ controllers/bet.js
@@ -1,4 +1,5 @@
 import { Bet } from '../models/bet.js';
+import Response from '../shared/response.js';
 
 const createBet = async (req, res, next) => {
   const bodyData = req.body;
@@ -9,17 +10,22 @@ const createBet = async (req, res, next) => {
 
 const deleteBet = async (req, res, next) => {
   const id = req.body.id;
+  const response = new Response();
   if (!id) {
-    return res.json('Please enter id');
+    return res.json(response.NOT_FOUND([]));
+  }
+
+  const existBet = await Bet.findOne({ _id: id });
+
+  if (!existBet) {
+    return res.json(response.NOT_FOUND([]));
   }
 
   try {
     const result = await Bet.deleteOne({ _id: id });
-    return res.json(result);
+    return res.json(response.OK([existBet]));
   } catch (error) {
-    res.json({
-      error,
-    });
+    return res.json(response.FAIL([]));
   }
 };
